# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Event Booking System API
  description: |
    REST API for an Event Booking System.

    Roles:
    - CUSTOMER: Can browse events and create/view their own bookings.
    - ORGANIZER: Can create, update, cancel, and manage their own events.

    Authentication:
    Most protected endpoints require a JWT obtained from the login endpoint.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new CUSTOMER or ORGANIZER account.
      operationId: register

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              customer:
                summary: Register as Customer
                value:
                  name: John Customer
                  email: customer@example.com
                  password: Password@123
                  role: CUSTOMER

              organizer:
                summary: Register as Organizer
                value:
                  name: Jane Organizer
                  email: organizer@example.com
                  password: Password@123
                  role: ORGANIZER

      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9.demo.jwt.token
                userId: 1
                name: John Customer
                role: CUSTOMER

        '400':
          description: Invalid request or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                message: Validation failed

        '409':
          description: User already exists
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticates a user and returns a JWT token.
      operationId: login

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: customer@example.com
              password: Password@123

      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9.demo.jwt.token
                userId: 1
                name: John Customer
                role: CUSTOMER

        '400':
          description: Invalid request or validation failure

        '401':
          description: Invalid email or password


  # ============================================================
  # EVENTS
  # ============================================================
...
  /api/events/{id}/bookings:
    get:
      tags:
        - Events
      summary: List bookings for an event
      description: |
        Returns bookings for an event owned by the authenticated organizer.

        Requires ORGANIZER role.
      operationId: getEventBookings
      security:
        - bearerAuth: []

      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'

      responses:
        '200':
          description: Event bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingPageResponse'
              example:
                items:
                  - id: 501
                    eventId: 101
                    customerId: 1
                    quantity: 5
                    status: CONFIRMED
                    createdAt: '2026-09-20T10:30:00Z'
                page: 0
                size: 20
                totalItems: 1
                totalPages: 1

        '401':
          description: Authentication required

        '403':
          description: ORGANIZER role required

        '404':
          description: Event not found
...
  /api/events/{eventId}/bookings:
    post:
      tags:
        - Bookings
      summary: Book tickets
      description: |
        Creates a booking for the authenticated customer.

        Requires CUSTOMER role.

        An optional Idempotency-Key can be supplied to safely retry
        the same booking request without creating duplicate bookings.
      operationId: createBooking
      security:
        - bearerAuth: []

      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event
          schema:
            type: integer
            format: int64
            example: 101

        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Unique key used to make the booking request idempotent.
            Reusing the same key prevents duplicate booking creation.
          schema:
            type: string
            example: booking-101-customer-1-001

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            example:
              quantity: 5

      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                id: 501
                eventId: 101
                customerId: 1
                quantity: 5
                status: CONFIRMED
                createdAt: '2026-09-20T10:30:00Z'

        '400':
          description: Invalid booking request

        '401':
          description: Authentication required

        '403':
          description: CUSTOMER role required

        '404':
          description: Event not found

        '409':
          description: Not enough tickets available or booking conflict
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: Get my bookings
      description: Returns bookings belonging to the authenticated customer.
      operationId: getMyBookings
      security:
        - bearerAuth: []

      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'

      responses:
        '200':
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingPageResponse'
              example:
                items:
                  - id: 501
                    eventId: 101
                    customerId: 1
                    quantity: 5
                    status: CONFIRMED
                    createdAt: '2026-09-20T10:30:00Z'
                page: 0
                size: 20
                totalItems: 1
                totalPages: 1

        '401':
          description: Authentication required

        '403':
          description: CUSTOMER role required
  /api/bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get my booking by ID
      description: Returns a booking belonging to the authenticated customer.
      operationId: getMyBooking

      security:
        - bearerAuth: []

      parameters:
        - name: id
          in: path
          required: true
          description: Booking ID
          schema:
            type: integer
            format: int64
            example: 501

      responses:
        '200':
          description: Booking retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                id: 501
                eventId: 101
                customerId: 1
                quantity: 5
                status: CONFIRMED
                createdAt: '2026-09-20T10:30:00Z'

        '401':
          description: Authentication required

        '403':
          description: CUSTOMER role required

        '404':
          description: Booking not found


# ==============================================================
# COMPONENTS
# ==============================================================
components:
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter the JWT token returned by /api/auth/login.

        Example:
        Bearer eyJhbGciOiJIUzI1NiJ9.demo.jwt.token


  # ------------------------------------------------------------
  # PARAMETERS
  # ------------------------------------------------------------
  parameters:
    EventId:
      name: id
      in: path
      required: true
      description: Event ID
      schema:
        type: integer
        format: int64
        example: 101
    Page:
      name: page
      in: query
      required: false
      description: Page number, starting from 0
      schema:
        type: integer
        minimum: 0
        default: 0
        example: 0
    Size:
      name: size
      in: query
      required: false
      description: Number of records per page
      schema:
        type: integer
        minimum: 1
        default: 20
        example: 20


  # ------------------------------------------------------------
  # SCHEMAS
  # ------------------------------------------------------------
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          example: John Customer
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: Password@123
        role:
          $ref: '#/components/schemas/Role'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: customer@example.com
        password:
          type: string
          format: password
          example: Password@123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9.demo.jwt.token
        userId:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: John Customer
        role:
          $ref: '#/components/schemas/Role'


    # ==========================================================
    # EVENT
    # ==========================================================
...
    BookingRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 10
          example: 5
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 501
        eventId:
          type: integer
          format: int64
          example: 101
        customerId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          example: 5
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
          format: date-time
          example: '2026-09-20T10:30:00Z'
    BookingStatus:
      type: string
      enum:
        - CONFIRMED
        - CANCELLED
      example: CONFIRMED


    # ==========================================================
    # PAGINATION
    # ==========================================================
...
    BookingPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalItems:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1


    # ==========================================================
    # USER
    # ==========================================================
    Role:
      type: string
      enum:
        - CUSTOMER
        - ORGANIZER
      example: CUSTOMER


    # ==========================================================
    # ERROR
    # ==========================================================
    ErrorResponse:
      type: object
      description: Generic error response. Actual fields depend on the application's exception handling.
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed